[refactor] Split Text into platform-specific files
