+++ Libraries/Text/Text.ios.js
@@ -1,0 +1,62 @@
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @providesModule Text
+ * @flow
+ */
+'use strict';
+
+const React = require('React');
+const TextBaseMixin = require('./TextBaseMixin.js');
+
+const createReactNativeComponentClass =
+  require('createReactNativeComponentClass');
+
+/**
+ * A React component for displaying text which supports nesting,
+ * styling, and touch handling.  In the following example, the nested title and
+ * body text will inherit the `fontFamily` from `styles.baseText`, but the title
+ * provides its own additional styles.  The title and body will stack on top of
+ * each other on account of the literal newlines:
+ *
+ * ```
+ * renderText: function() {
+ *   return (
+ *     <Text style={styles.baseText}>
+ *       <Text style={styles.titleText} onPress={this.onPressTitle}>
+ *         {this.state.titleText + '\n\n'}
+ *       </Text>
+ *       <Text numberOfLines={5}>
+ *         {this.state.bodyText}
+ *       </Text>
+ *     </Text>
+ *   );
+ * },
+ * ...
+ * var styles = StyleSheet.create({
+ *   baseText: {
+ *     fontFamily: 'Cochin',
+ *   },
+ *   titleText: {
+ *     fontSize: 20,
+ *     fontWeight: 'bold',
+ *   },
+ * };
+ * ```
+ */
+const Text = React.createClass({
+  mixins: [TextBaseMixin],
+  render(): ReactElement {
+    const newProps = this.getNewProps();
+    return <RCTText {...newProps} />;
+  },
+});
+
+const RCTText = createReactNativeComponentClass(TextBaseMixin.viewConfig);
+
+module.exports = Text;
